% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2tfdraft.R
\name{convertTS}
\alias{convertTS}
\title{convert to ts
take model and basic sequence to create time series}
\usage{
convertTS(gco, originalData, model = NULL, modelFunction = NULL,
  target = NULL, ...)
}
\arguments{
\item{gco}{a basic sequence}

\item{model}{a model}
}
\description{
convert to ts
take model and basic sequence to create time series
}
