% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2tfdraft.R
\name{genGreySeq}
\alias{genGreySeq}
\title{genGreySeq
generate base object for others function}
\usage{
genGreySeq(valstep = 9, minval = NULL, maxval = NULL, sourceData = NULL,
  style = "SD")
}
\arguments{
\item{valstep}{vector or numeric indicate level of each feature}

\item{sourceData}{orginal data of the model}
}
\description{
genGreySeq
generate base object for others function
}
\author{
Pj.
}
